Former-commit-id: 7787bce444c107a21c8f0633bc58d8985e77f25c

+++ backend/src/main/java/com/ssafy/api/model/Friendship.java
@@ -26,12 +26,12 @@ public class Friendship {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "ID")
-	private int ID;
+	private long ID;
 
 	@Column(name = "FRIEND1_ID")
-	private int friend1ID;
+	private long friend1ID;
 
 	@Column(name = "FRIEND2_ID")
-	private int friend2ID;
+	private long friend2ID;
 
 }
